refactor Protocol enum
