+++ src/main/java/org/rocstreaming/roctoolkit/ProtocolUtils.java
@@ -1,0 +1,17 @@
+package org.rocstreaming.roctoolkit;
+
+import lombok.NoArgsConstructor;
+
+@SuppressWarnings("unused") // used by JNI
+@NoArgsConstructor
+class ProtocolUtils {
+
+    private static Protocol getByValue(int value) {
+        for (Protocol protocol : Protocol.values()) {
+            if (value == protocol.value) {
+                return protocol;
+            }
+        }
+        return null;
+    }
+}
